move methods into a concern
