+++ spec/models/appeal_spec.rb
@@ -12,14 +12,16 @@ describe Appeal do
     Timecop.freeze(Time.utc(2019, 1, 1, 12, 0, 0))
   end
 
-  context "#tasks_structure" do
-    let!(:root_task) { FactoryBot.create(:root_task, appeal: appeal) }
+  context "includes PrintsTaskTree concern" do
+    context "#structure" do
+      let!(:root_task) { FactoryBot.create(:root_task, appeal: appeal) }
 
-    subject { appeal.structure(:id) }
+      subject { appeal.structure(:id) }
 
-    it "returns the task structure" do
-      expect_any_instance_of(RootTask).to receive(:structure).with(:id)
-      expect(subject.key?(:"Appeal #{appeal.id}")).to be_truthy
+      it "returns the task structure" do
+        expect_any_instance_of(RootTask).to receive(:structure).with(:id)
+        expect(subject.key?(:"Appeal #{appeal.id}")).to be_truthy
+      end
     end
   end
