+++ spec/models/legacy_appeal_spec.rb
@@ -13,25 +13,27 @@ describe LegacyAppeal do
     create(:legacy_appeal, vacols_case: vacols_case)
   end
 
-  context "#tasks_structure" do
-    let!(:root_task) { FactoryBot.create(:root_task, appeal: appeal) }
-    let(:vacols_case) { create(:case, bfcorlid: "123456789S") }
+  context "includes PrintsTaskTree concern" do
+    context "#structure" do
+      let!(:root_task) { FactoryBot.create(:root_task, appeal: appeal) }
+      let(:vacols_case) { create(:case, bfcorlid: "123456789S") }
 
-    subject { appeal.structure(:id) }
+      subject { appeal.structure(:id) }
 
-    it "returns the task structure" do
-      expect_any_instance_of(RootTask).to receive(:structure).with(:id)
-      expect(subject.key?(:"LegacyAppeal #{appeal.id}")).to be_truthy
-    end
-
-    context "the appeal has more than one parentless task" do
-      let!(:colocated_task) { FactoryBot.create(:colocated_task, appeal: appeal, parent: nil) }
-
-      it "returns all parentless tasks" do
+      it "returns the task structure" do
         expect_any_instance_of(RootTask).to receive(:structure).with(:id)
-        expect_any_instance_of(ColocatedTask).to receive(:structure).with(:id)
         expect(subject.key?(:"LegacyAppeal #{appeal.id}")).to be_truthy
-        expect(subject[:"LegacyAppeal #{appeal.id}"].count).to eq 2
+      end
+
+      context "the appeal has more than one parentless task" do
+        let!(:colocated_task) { FactoryBot.create(:colocated_task, appeal: appeal, parent: nil) }
+
+        it "returns all parentless tasks" do
+          expect_any_instance_of(RootTask).to receive(:structure).with(:id)
+          expect_any_instance_of(ColocatedTask).to receive(:structure).with(:id)
+          expect(subject.key?(:"LegacyAppeal #{appeal.id}")).to be_truthy
+          expect(subject[:"LegacyAppeal #{appeal.id}"].count).to eq 2
+        end
       end
     end
   end
