+++ spec/models/task_spec.rb
@@ -1,36 +1,38 @@
 # frozen_string_literal: true
 
 describe Task do
-  describe ".structure" do
-    let(:root_task) { FactoryBot.create(:root_task, :on_hold) }
-    let!(:bva_task) { FactoryBot.create(:bva_dispatch_task, :in_progress, parent: root_task) }
-    let(:judge_task) { FactoryBot.create(:ama_judge_task, :completed, parent: root_task) }
-    let!(:attorney_task) { FactoryBot.create(:ama_attorney_task, :completed, parent: judge_task) }
+  context "includes PrintsTaskTree concern" do
+    describe ".structure" do
+      let(:root_task) { FactoryBot.create(:root_task, :on_hold) }
+      let!(:bva_task) { FactoryBot.create(:bva_dispatch_task, :in_progress, parent: root_task) }
+      let(:judge_task) { FactoryBot.create(:ama_judge_task, :completed, parent: root_task) }
+      let!(:attorney_task) { FactoryBot.create(:ama_attorney_task, :completed, parent: judge_task) }
 
-    subject { root_task.structure(:id, :status) }
+      subject { root_task.structure(:id, :status) }
 
-    it "outputs the task structure" do
-      root_key = "#{root_task.class.name} #{root_task.id}, #{root_task.status}".to_sym
-      judge_key = "#{judge_task.class.name} #{judge_task.id}, #{judge_task.status}".to_sym
-      bva_key = "#{bva_task.class.name} #{bva_task.id}, #{bva_task.status}".to_sym
-      attorney_key = "#{attorney_task.class.name} #{attorney_task.id}, #{attorney_task.status}".to_sym
+      it "outputs the task structure" do
+        root_key = "#{root_task.class.name} #{root_task.id}, #{root_task.status}".to_sym
+        judge_key = "#{judge_task.class.name} #{judge_task.id}, #{judge_task.status}".to_sym
+        bva_key = "#{bva_task.class.name} #{bva_task.id}, #{bva_task.status}".to_sym
+        attorney_key = "#{attorney_task.class.name} #{attorney_task.id}, #{attorney_task.status}".to_sym
 
-      expect(subject.key?(root_key)).to be_truthy
-      expect(subject[root_key].count).to eq 2
-      judge_task_found = false
-      bva_task_found = false
-      subject[root_key].each do |child_task|
-        if child_task.key? judge_key
-          judge_task_found = true
-          expect(child_task[judge_key].count).to eq 1
-          expect(child_task[judge_key].first.key?(attorney_key)).to be_truthy
-          expect(child_task[judge_key].first[attorney_key]).to eq []
-        elsif child_task.key? bva_key
-          bva_task_found = true
+        expect(subject.key?(root_key)).to be_truthy
+        expect(subject[root_key].count).to eq 2
+        judge_task_found = false
+        bva_task_found = false
+        subject[root_key].each do |child_task|
+          if child_task.key? judge_key
+            judge_task_found = true
+            expect(child_task[judge_key].count).to eq 1
+            expect(child_task[judge_key].first.key?(attorney_key)).to be_truthy
+            expect(child_task[judge_key].first[attorney_key]).to eq []
+          elsif child_task.key? bva_key
+            bva_task_found = true
+          end
         end
+        expect(judge_task_found).to be_truthy
+        expect(bva_task_found).to be_truthy
       end
-      expect(judge_task_found).to be_truthy
-      expect(bva_task_found).to be_truthy
     end
   end
